Organize component folder
